+++ src/app/services/page.tsx
@@ -1,4 +1,4 @@
-import Services from "@/components/unauthorizePage/Services";
+import OurServices from "@/components/service/OurServices";
 import authOptions from "@/lib/auth/options";
 import { getServerSession } from "next-auth";
 
@@ -7,7 +7,7 @@ export default async function Page() {
     const token = session?.user?.token;*/
     return (
         <div className="mx-auto">
-            <Services />
+            <OurServices />
         </div>
     );
 }
